fix incorrect typecast in features/tv (re: c250bcb2)
The new typecasts in the ST_ATIME_NSEC_GET macro in
src/lib/libast/features/tv, like this one:

  #define ST_ATIME_NSEC_GET(st)   ((uint32_t)(st)->st_atim.tv_nsec)

are not correct as ST_ATIME_NSEC_GET is also used as an lvalue:

54:if ( !no_stat_nsec ) {
55:	#define ST_ATIME_NSEC_SET(st,n)	(ST_ATIME_NSEC_GET(st)=(n))
56:	#define ST_CTIME_NSEC_SET(st,n)	(ST_CTIME_NSEC_GET(st)=(n))
57:	#define ST_MTIME_NSEC_SET(st,n)	(ST_MTIME_NSEC_GET(st)=(n))
58:}
59:endif

which, when used, causes a compilation failure.

Remove the typecast from the ST_ATIME_NSEC_GET and insert other
typecasts elsewhere as necesssary to silence compiler warnings (at
least on clang on macOS; warnings now popping up on other systems
may need to be corrected later).

+++ src/cmd/ksh93/sh/parse.c
@@ -1006,7 +1006,7 @@ static Shnode_t *funct(Lex_t *lexp)
 	}
 	sh.st.staklist = (struct slnod*)slp;
 	last = fctell();
-	fp->functline = (last-first);
+	fp->functline = (int)(last-first);
 	fp->functtre = t;
 	sh.mktype = in_mktype;
 	if(sh.funlog)
